refactor: user service routes
