+++ api/src/services/user/handlers/addTeacher/index.js
@@ -1,0 +1,4 @@
+import { addTeacherHandler } from './handler';
+import { addTeacherOptions } from './options';
+
+export { addTeacherHandler, addTeacherOptions };
