+++ api/src/services/user/handlers/deleteUser/index.js
@@ -1,0 +1,4 @@
+import { deleteUserHandler } from './handler';
+import { deleteUserOptions } from './options';
+
+export { deleteUserHandler, deleteUserOptions };
