+++ api/src/services/user/handlers/getAllUsers/index.js
@@ -1,0 +1,4 @@
+import { getAllUsersHandler } from './handler';
+import { getAllUsersOptions } from './options';
+
+export { getAllUsersHandler, getAllUsersOptions };
