+++ api/src/services/user/handlers/getAllUsers/options.js
@@ -1,0 +1,11 @@
+export const getAllUsersOptions = {
+  schema: {
+    response: {
+      '4xx': { $ref: '4xx#' },
+      '5xx': { $ref: '5xx#' },
+    },
+  },
+  async onRequest(req) {
+    await this.auth({ req, isAdminOnly: true });
+  },
+};
