+++ api/src/services/user/handlers/refreshToken/index.js
@@ -1,0 +1,4 @@
+import { refreshTokenHandler } from './handler';
+import { refreshTokenOptions } from './options';
+
+export { refreshTokenHandler, refreshTokenOptions };
