+++ api/src/services/user/handlers/refreshToken/options.js
@@ -1,0 +1,15 @@
+export const refreshTokenOptions = {
+  schema: {
+    body: {
+      type: 'object',
+      properties: {
+        refreshToken: { type: 'string' },
+      },
+      required: ['refreshToken'],
+    },
+    response: {
+      '4xx': { $ref: '4xx#' },
+      '5xx': { $ref: '5xx#' },
+    },
+  },
+};
