+++ api/src/services/user/handlers/removeTeacher/options.js
@@ -1,0 +1,18 @@
+export const removeTeacherOptions = {
+  schema: {
+    body: {
+      type: 'object',
+      properties: {
+        id: { type: 'number' },
+      },
+      required: ['id'],
+    },
+    response: {
+      '4xx': { $ref: '4xx#' },
+      '5xx': { $ref: '5xx#' },
+    },
+  },
+  async onRequest(req) {
+    await this.auth({ req, isAdminOnly: true });
+  },
+};
