+++ api/src/services/user/handlers/self/index.js
@@ -1,0 +1,4 @@
+import { selfHandler } from './handler';
+import { selfOptions } from './options';
+
+export { selfHandler, selfOptions };
