+++ api/src/services/user/handlers/signIn/index.js
@@ -1,0 +1,4 @@
+import { signInHandler } from './handler';
+import { signInOptions } from './options';
+
+export { signInHandler, signInOptions };
