+++ api/src/services/user/handlers/signUp/index.js
@@ -1,0 +1,4 @@
+import { signUpHandler } from './handler';
+import { signUpOptions } from './options';
+
+export { signUpHandler, signUpOptions };
