+++ api/src/services/user/handlers/updateUser/index.js
@@ -1,0 +1,4 @@
+import { updateUserHandler } from './handler';
+import { updateUserOptions } from './options';
+
+export { updateUserHandler, updateUserOptions };
