+++ api/src/services/user/handlers/updateUser/options.js
@@ -1,0 +1,32 @@
+export const updateUserOptions = {
+  schema: {
+    body: {
+      type: 'object',
+      properties: {
+        email: { type: 'string' },
+        password: { type: 'string', pattern: '^(?=.*\\d)(?=.*\\D).{5,}$' },
+        firstName: { type: 'string' },
+        lastName: { type: 'string' },
+        isConfirmed: { type: 'boolean' },
+        isSuperAdmin: { type: 'boolean' },
+      },
+    },
+    params: {
+      type: 'object',
+      properties: {
+        userId: { type: 'number' },
+      },
+      required: ['userId'],
+    },
+    response: {
+      '4xx': { $ref: '4xx#' },
+      '5xx': { $ref: '5xx#' },
+    },
+  },
+  async onRequest(req) {
+    await this.auth({ req, isAdminOnly: true });
+  },
+  async preHandler({ body: { email } }) {
+    this.validateEmail({ email });
+  },
+};
